Split NDJSON byte scanner
